feat: support new config system
