+++ src/configs/style.ts
@@ -1,0 +1,22 @@
+import globals from '../globals.json';
+import jest from '../index';
+import legacy from '../legacy';
+
+export default [
+  {
+    files: ['**/*.snap'],
+    plugins: {
+      jest,
+    },
+    processor: 'jest/.snap',
+  },
+  {
+    languageOptions: {
+      globals,
+    },
+    plugins: {
+      jest,
+    },
+    rules: legacy.configs.style.rules,
+  },
+];
