+++ src/index.ts
@@ -1,87 +1,6 @@
-import { readdirSync } from 'fs';
-import { join, parse } from 'path';
-import type { TSESLint } from '@typescript-eslint/utils';
-import globals from './globals.json';
-import * as snapshotProcessor from './processors/snapshot-processor';
+import legacy from './legacy';
 
-type RuleModule = TSESLint.RuleModule<string, unknown[]> & {
-  meta: Required<Pick<TSESLint.RuleMetaData<string>, 'docs'>>;
-};
-
-// v5 of `@typescript-eslint/experimental-utils` removed this
-declare module '@typescript-eslint/utils/dist/ts-eslint/Rule' {
-  export interface RuleMetaDataDocs {
-    category: 'Best Practices' | 'Possible Errors';
-  }
-}
-
-// copied from https://github.com/babel/babel/blob/d8da63c929f2d28c401571e2a43166678c555bc4/packages/babel-helpers/src/helpers.js#L602-L606
-/* istanbul ignore next */
-const interopRequireDefault = (obj: any): { default: any } =>
-  obj && obj.__esModule ? obj : { default: obj };
-
-const importDefault = (moduleName: string) =>
-  // eslint-disable-next-line @typescript-eslint/no-require-imports
-  interopRequireDefault(require(moduleName)).default;
-
-const rulesDir = join(__dirname, 'rules');
-const excludedFiles = ['__tests__', 'detectJestVersion', 'utils'];
-
-const rules = readdirSync(rulesDir)
-  .map(rule => parse(rule).name)
-  .filter(rule => !excludedFiles.includes(rule))
-  .reduce<Record<string, RuleModule>>(
-    (acc, curr) => ({
-      ...acc,
-      [curr]: importDefault(join(rulesDir, curr)) as RuleModule,
-    }),
-    {},
-  );
-
-const recommendedRules = Object.entries(rules)
-  .filter(([, rule]) => rule.meta.docs.recommended)
-  .reduce(
-    (acc, [name, rule]) => ({
-      ...acc,
-      [`jest/${name}`]: rule.meta.docs.recommended,
-    }),
-    {},
-  );
-
-const allRules = Object.entries(rules)
-  .filter(([, rule]) => !rule.meta.deprecated)
-  .reduce(
-    (acc, [name]) => ({
-      ...acc,
-      [`jest/${name}`]: 'error',
-    }),
-    {},
-  );
-
-const createConfig = (rules: Record<string, TSESLint.Linter.RuleLevel>) => ({
-  plugins: ['jest'],
-  env: { 'jest/globals': true },
-  rules,
-});
-
-export = {
-  configs: {
-    all: createConfig(allRules),
-    recommended: createConfig(recommendedRules),
-    style: createConfig({
-      'jest/no-alias-methods': 'warn',
-      'jest/prefer-to-be': 'error',
-      'jest/prefer-to-contain': 'error',
-      'jest/prefer-to-have-length': 'error',
-    }),
-  },
-  environments: {
-    globals: {
-      globals,
-    },
-  },
-  processors: {
-    '.snap': snapshotProcessor,
-  },
-  rules,
+export default {
+  rules: legacy.rules,
+  processors: legacy.processors,
 };
